// src/main/services/http-client-types.ts
/*
 * Copyright 2026, gematik GmbH
 *
 * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
 * European Commission – subsequent versions of the EUPL (the "Licence").
 * You may not use this work except in compliance with the Licence.
 *
 * You find a copy of the Licence in the "Licence" file or at
 * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the Licence is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expressed or implied.
 * In case of changes by gematik find details in the "Readme" file.
 *
 * See the Licence for the specific language governing permissions and limitations under the Licence.
 *
 * ******
 *
 * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
 */

import { IncomingHttpHeaders } from 'http';

export type HTTPClientConfig = {
  headers?: Record<string, string | string[]>;
  body?: string;
  json?: Record<string, any>;
  timeout?: number;
  followRedirect?: boolean;
  useProxyForConnector?: boolean;
  username?: string;
  password?: string;
  ciphers?: string;
  https?: {
    rejectUnauthorized?: boolean;
    certificate?: string | Buffer | (string | Buffer)[];
    key?: string | Buffer | (string | Buffer)[];
    certificateAuthority?: string | Buffer | (string | Buffer)[];
    pfx?: Buffer;
    passphrase?: string;
    checkServerIdentity?: unknown;
  };
};

export enum HTTP_METHODS {
  POST,
  GET,
}

export type TClientRes = {
  data: any;
  status: number;
  headers?: IncomingHttpHeaders;
};